Add files via upload
The purpose of this code was to reinforce how to use Kmeans and understand the context of the clusters.  I wanted to reinforce using clusters in a different context which allowed me to understand what type of variables can be used as well as having to normalize different values.  I also wanted to know how to visualize the clusters to look into the context of the clusters.  The new code thing that I did was a ggarrange of plotting 2 graphs next to each other.  I had to look up how to evaluate the number of clusters that were optimal after/during class after listening to the discussion.