sg4sgz file upload
The purpose of this code is to reinforce concepts learned throughout this semester, including but limited to clustering, kmeans, data cleaning, and data visualization. I applied all four of those ideas in my code to gain a better understanding of those topics. The dataset was unfamiliar at first so it was useful to understand the data in order to perform further operations on the dataset and be able to construct an analysis on it.
A figure generated from my code of the final data analysis is attached in the folder as sg4sgz.png. There are additional figures and analysis that is featured in the individual Rmarkdown file (sg4sgz.Rmd).